+++ site/package.json
@@ -8,11 +8,12 @@
     "gatsby-plugin-google-analytics": "^2.0.7",
     "gatsby-plugin-react-helmet": "^3.0.0",
     "gatsby-plugin-styled-components": "^3.0.0",
-    "react": "^16.7.0-alpha.0",
-    "react-dom": "^16.7.0-alpha.0",
+    "react": "^16.8.1",
+    "react-dom": "^16.8.1",
     "react-helmet": "^5.2.0",
     "react-highlight-words": "^0.16.0",
-    "styled-components": "^4.0.2"
+    "styled-components": "^4.0.2",
+    "tmp-cache": "^1.0.0"
   },
   "license": "MIT",
   "scripts": {
